+++ packages/components/popover/src/free-solo-popover.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import {DismissButton, Overlay} from "@react-aria/overlays";
-import {HTMLMotionProps, motion} from "framer-motion";
+import {domAnimation, HTMLMotionProps, LazyMotion, m} from "framer-motion";
 import {mergeProps} from "@react-aria/utils";
 import {getTransformOrigins} from "@tw-material/aria-utils";
 import {TRANSITION_VARIANTS} from "@tw-material/framer-transitions";
@@ -30,19 +30,21 @@ const FreeSoloPopoverWrapper = ({
   return disableAnimation ? (
     <div {...otherProps}>{children}</div>
   ) : (
-    <motion.div
-      animate="enter"
-      exit="exit"
-      initial="initial"
-      style={{
-        ...style,
-        ...getTransformOrigins(placement === "center" ? "top" : placement),
-      }}
-      variants={TRANSITION_VARIANTS.scaleSpringOpacity}
-      {...mergeProps(otherProps, motionProps)}
-    >
-      {children}
-    </motion.div>
+    <LazyMotion features={domAnimation}>
+      <m.div
+        animate="enter"
+        exit="exit"
+        initial="initial"
+        style={{
+          ...style,
+          ...getTransformOrigins(placement === "center" ? "top" : placement),
+        }}
+        variants={TRANSITION_VARIANTS.scaleSpringOpacity}
+        {...mergeProps(otherProps, motionProps)}
+      >
+        {children}
+      </m.div>
+    </LazyMotion>
   );
 };
 
@@ -63,8 +65,6 @@ const FreeSoloPopover = forwardRef<"div", FreeSoloPopoverProps>((props, ref) => 
     getContentProps,
   } = usePopover({
     ...props,
-    // avoid closing the popover when navigating with the keyboard
-    shouldCloseOnInteractOutside: undefined,
     ref,
   });
 
@@ -78,13 +78,15 @@ const FreeSoloPopover = forwardRef<"div", FreeSoloPopoverProps>((props, ref) => 
     }
 
     return (
-      <motion.div
-        animate="enter"
-        exit="exit"
-        initial="exit"
-        variants={TRANSITION_VARIANTS.fade}
-        {...(getBackdropProps() as HTMLMotionProps<"div">)}
-      />
+      <LazyMotion features={domAnimation}>
+        <m.div
+          animate="enter"
+          exit="exit"
+          initial="exit"
+          variants={TRANSITION_VARIANTS.fade}
+          {...(getBackdropProps() as HTMLMotionProps<"div">)}
+        />
+      </LazyMotion>
     );
   }, [backdrop, disableAnimation, getBackdropProps]);