+++ packages/components/scroll-shadow/src/scroll-shadow.tsx
@@ -1,0 +1,15 @@
+import {forwardRef} from "@tw-material/system";
+
+import {UseScrollShadowProps, useScrollShadow} from "./use-scroll-shadow";
+
+export interface ScrollShadowProps extends UseScrollShadowProps {}
+
+const ScrollShadow = forwardRef<"div", ScrollShadowProps>((props, ref) => {
+  const {Component, children, getBaseProps} = useScrollShadow({...props, ref});
+
+  return <Component {...getBaseProps()}>{children}</Component>;
+});
+
+ScrollShadow.displayName = "ScrollShadow";
+
+export default ScrollShadow;