+++ packages/theme/src/plugin/index.ts
@@ -16,6 +16,26 @@ export const material3 = (options: Options) => {
   return plugin(
     ({addUtilities, addVariant}) => {
       addUtilities(resolved.utilities);
+      addUtilities({
+        ".scrollbar-hide": {
+          "-ms-overflow-style": "none",
+
+          "scrollbar-width": "none",
+
+          "&::-webkit-scrollbar": {
+            display: "none",
+          },
+        },
+        ".scrollbar-default": {
+          "-ms-overflow-style": "auto",
+
+          "scrollbar-width": "auto",
+
+          "&::-webkit-scrollbar": {
+            display: "block",
+          },
+        },
+      });
       resolved.variants.forEach(({name, definition}) => addVariant(name, definition));
     },
     {