+++ packages/components/listbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tw-material/listbox
 
+## 1.0.3
+
+### Patch Changes
+
+- update
+  - @tw-material/divider@1.0.1
+
 ## 1.0.2
 
 ### Patch Changes