+++ packages/components/slider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tw-material/slider
 
+## 1.0.8
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @tw-material/ripple@1.1.3
+  - @tw-material/theme@1.1.7
+  - @tw-material/tooltip@1.0.7
+
 ## 1.0.7
 
 ### Patch Changes